REFACTOR: rename es2019 to es2015
