+++ addon/connect.js
@@ -1,6 +1,6 @@
 import Component from '@ember/component';
 import wrapEmberObject from './ember-object';
-import wrapEs2019Class from './es2019-class';
+import wrapEs2015Class from './es2015-class';
 
 export default (stateToComputed, dispatchToActions=() => ({})) => {
   return IncomingComponent => {
@@ -10,6 +10,6 @@ export default (stateToComputed, dispatchToActions=() => ({})) => {
       return wrapEmberObject(stateToComputed, dispatchToActions, WrappedComponent);
     }
 
-    return wrapEs2019Class(stateToComputed, dispatchToActions, WrappedComponent);
+    return wrapEs2015Class(stateToComputed, dispatchToActions, WrappedComponent);
   };
 };
