+++ addon/es2015-class.js
