+++ tests/dummy/app/components/basic-init/component.js
@@ -2,11 +2,9 @@ import Component from '@ember/component';
 import { get } from '@ember/object';
 import { connect } from 'ember-redux';
 
-var stateToComputed = function(state) {
-  return {
-    low: state.low
-  };
-};
+var stateToComputed = state => ({
+  low: state.low
+});
 
 var ConnectComponent = Component.extend({
