+++ tests/dummy/app/components/class-init/template.hbs
