--- tests/dummy/app/es19/template.hbs
@@ -1,1 +1,0 @@
-{{es-19}}
