+++ tests/dummy/app/es2015/template.hbs
@@ -1,0 +1,1 @@
+{{class-init}}
